Security Update
Previously the database file was unencrypted, during use, which is not good, since it makes it so that anyone can just read the file, now the file is never unencrypted, and only a variable containing the file data is, and when you are done, the program encrypts the variable, then it writes the variables encrypted data to the database file. and thus, the Database file always stays encrypted.